+++ src/Certes/IKey.cs
@@ -1,8 +1,18 @@
-﻿using System.Text;
+﻿using System;
+using System.IO;
+using System.Text;
+using Certes.Acme.Resource;
+using Certes.Crypto;
 using Certes.Json;
 using Certes.Jws;
 using Newtonsoft.Json;
+using Org.BouncyCastle.Asn1;
+using Org.BouncyCastle.Asn1.X509;
+using Org.BouncyCastle.Crypto.Operators;
+using Org.BouncyCastle.Math;
+using Org.BouncyCastle.OpenSsl;
 using Org.BouncyCastle.Security;
+using Org.BouncyCastle.X509;
 
 namespace Certes
 {
@@ -17,7 +27,7 @@ namespace Certes
         /// <value>
         /// The algorithm.
         /// </value>
-        KeyAlgorithm Algorithm { get;}
+        KeyAlgorithm Algorithm { get; }
 
         /// <summary>
         /// Gets the json web key.
@@ -33,6 +43,8 @@ namespace Certes
     /// </summary>
     public static class ISignatureKeyExtensions
     {
+        private static readonly DerObjectIdentifier acmeValidationV1Id = new DerObjectIdentifier("1.3.6.1.5.5.7.1.30.1");
+        private static readonly KeyAlgorithmProvider signatureAlgorithmProvider = new KeyAlgorithmProvider();
         private static readonly JsonSerializerSettings thumbprintSettings = JsonUtil.CreateSettings();
 
         /// <summary>
@@ -66,7 +78,7 @@ namespace Certes
         /// </summary>
         /// <param name="key">The key.</param>
         /// <param name="token">The challenge token.</param>
-        /// <returns></returns>
+        /// <returns>The key authorization string.</returns>
         public static string KeyAuthorization(this IKey key, string token)
         {
             var jwkThumbprintEncoded = key.Thumbprint();
@@ -78,12 +90,59 @@ namespace Certes
         /// </summary>
         /// <param name="key">The key.</param>
         /// <param name="token">The challenge token.</param>
-        /// <returns></returns>
+        /// <returns>The DNS text value for dns-01 validation.</returns>
         public static string DnsTxt(this IKey key, string token)
         {
             var keyAuthz = key.KeyAuthorization(token);
             var hashed = DigestUtilities.CalculateDigest("SHA256", Encoding.UTF8.GetBytes(keyAuthz));
             return JwsConvert.ToBase64String(hashed);
+        }
+
+        /// <summary>
+        /// Generates the certificate for <see cref="ChallengeTypes.TlsAlpn01" /> validation.
+        /// </summary>
+        /// <param name="key">The key.</param>
+        /// <param name="token">The <see cref="ChallengeTypes.TlsAlpn01" /> token.</param>
+        /// <param name="subjectName">Name of the subject.</param>
+        /// <param name="certificateKey">The certificate key pair.</param>
+        /// <returns>The tls-alpn-01 certificate in PEM.</returns>
+        public static string TlsAlpnCertificate(this IKey key, string token, string subjectName, IKey certificateKey)
+        {
+            var keyAuthz = key.KeyAuthorization(token);
+            var hashed = DigestUtilities.CalculateDigest("SHA256", Encoding.UTF8.GetBytes(keyAuthz));
+
+            var (_, keyPair) = signatureAlgorithmProvider.GetKeyPair(certificateKey.ToDer());
+
+            var signatureFactory = new Asn1SignatureFactory(certificateKey.Algorithm.ToPkcsObjectId(), keyPair.Private, new SecureRandom());
+            var gen = new X509V3CertificateGenerator();
+            var certName = new X509Name($"CN={subjectName}");
+            var serialNo = BigInteger.ProbablePrime(120, new SecureRandom());
+
+            gen.SetSerialNumber(serialNo);
+            gen.SetSubjectDN(certName);
+            gen.SetIssuerDN(certName);
+            gen.SetNotBefore(DateTime.UtcNow);
+            gen.SetNotAfter(DateTime.UtcNow.AddDays(7));
+            gen.SetPublicKey(keyPair.Public);
+
+            // SAN for validation
+            var gns = new[] { new GeneralName(GeneralName.DnsName, subjectName) };
+            gen.AddExtension(X509Extensions.SubjectAlternativeName.Id, false, new GeneralNames(gns));
+
+            // ACME-TLS/1
+            gen.AddExtension(
+                acmeValidationV1Id,
+                true,
+                hashed);
+
+            var newCert = gen.Generate(signatureFactory);
+
+            using (var sr = new StringWriter())
+            {
+                var pemWriter = new PemWriter(sr);
+                pemWriter.WriteObject(newCert);
+                return sr.ToString();
+            }
         }
     }
 }