+++ docs/CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Added
+- Support `tls-alpn-01` challenge ([#125][i125])
 
 ## [2.2.2] - 2018-05-31
 ### Changed
@@ -93,3 +95,4 @@ All notable changes to this project will be documented in this file.
 [i100]: https://github.com/fszlin/certes/issues/100
 [i106]: https://github.com/fszlin/certes/issues/106
 [i112]: https://github.com/fszlin/certes/issues/112
+[i125]: https://github.com/fszlin/certes/issues/125